apps/auth-sms/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-jose==3.3.0
pydantic==2.9.2
phonenumbers==8.13.45